Auto stash before merge of "master" and "origin/master"

+++ save_my_flag/save_my_flag/player.h
@@ -13,8 +13,8 @@ public:
 	int getRadious();
 	bool getIsDead();
 
-	int posX;
-	int posY;
+	float posX;
+	float posY;
 	int radious;
 	int width;
 	int height;
